pyproject.toml
[project]
name = "kizaru"
version = "2.1"
description = "Add your description here"
dependencies = [
    "pigpio>=1.78",
    "gpiozero>=2.0",
    "psutil>=5.9.6",
    "loguru>=0.7.2",
    "notifiers==1.3.3",
    "line-bot-sdk>=3.5.1",
    "python-dotenv>=1.0.0",
    "xdg-base-dirs>=6.0.1",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0.post1",
]
readme = "README.md"
requires-python = ">= 3.10"

#[project.scripts]
#main = 'kizaru.main:main'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
init = { cmd = "cp -n src/kizaru/secret_example.py src/kizaru/secret.py && cp -n .env.example .env" }
main = { cmd = "python main.py" }
dev_local = { cmd = "python main.py dev_local" }
# Pueue commands (need pueue) wrapped by python thin script
# local command is called by "$command local" (without "daemon_local").
pueue = { cmd = "python daemon_scripts.py" }
daemon = { cmd = "python daemon_scripts.py add 'python main.py'" }
stop = { cmd = "python daemon_scripts.py stop" }
daemon_local = { cmd = "python daemon_scripts.py add local 'python main.py dev_local'" }

[tool.hatch.metadata]
allow-direct-references = true
